innovation_project/ImageProcessingUtils/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aspose-3d==23.7.0
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.43
botocore==1.31.43
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
einops==0.6.1
encodec==0.1.1
filelock==3.12.3
frozenlist==1.3.3
fsspec==2023.9.0
funcy==2.0
google-search-results==2.4.2
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
jmespath==1.0.1
langchain==0.0.147
lazy_loader==0.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.24.1
numpy-stl==3.0.0
openai==0.27.4
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
packaging==23.1
Pillow==9.4.0
pydantic==1.10.7
PySimpleGUI==4.60.4
python-dateutil==2.8.2
python-utils==3.7.0
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.8.8
requests==2.28.2
s3transfer==0.6.2
safetensors==0.3.3
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
SQLAlchemy==1.4.47
suno-bark @ git+https://github.com/suno-ai/bark.git@cb89688307c28cbd2d8bbfc78e534b9812673a26
sympy==1.12
tenacity==8.2.2
tifffile==2023.8.12
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
tqdm==4.65.0
transformers==4.33.1
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.1
